Initial scaffolding changes and HealthChecks.UI.SqlServer.Storage package
