+++ AspNetCore.Diagnostics.HealthChecks.sln
@@ -58,15 +58,12 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.UI.Sample", "samples\HealthChecks.UI.Sample\HealthChecks.UI.Sample.csproj", "{7C8F7A48-2FF9-4CF4-AAAC-142053E34FC0}"
 	ProjectSection(ProjectDependencies) = postProject
 		{B7C5C213-677D-4B03-ADA9-101299BCA047} = {B7C5C213-677D-4B03-ADA9-101299BCA047}
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6} = {B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}
 		{91DF8A73-8174-4EA2-ABFA-2557B71EE090} = {91DF8A73-8174-4EA2-ABFA-2557B71EE090}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Publisher.ApplicationInsights", "src\HealthChecks.Publisher.ApplicationInsights\HealthChecks.Publisher.ApplicationInsights.csproj", "{EEF1E206-DEE3-45F2-A753-73D2FFFF8EDA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Sample", "samples\HealthChecks.Sample\HealthChecks.Sample.csproj", "{0A7AF6F0-DEA6-4195-91F3-CD9E2484E1F3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.UI.Client", "src\HealthChecks.UI.Client\HealthChecks.UI.Client.csproj", "{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Publisher.Prometheus", "src\HealthChecks.Publisher.Prometheus\HealthChecks.Publisher.Prometheus.csproj", "{AA04C349-6402-4241-A611-60EA3CFBD4ED}"
 EndProject
@@ -117,6 +114,12 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docker-image", "docker-image", "{95119F6F-87C8-45B8-8D95-61736FBEBEDE}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Solr", "src\HealthChecks.Solr\HealthChecks.Solr.csproj", "{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HealthChecks.UI.SqlServer.Storage", "src\HealthChecks.UI.SqlServer.Storage\HealthChecks.UI.SqlServer.Storage.csproj", "{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HealthChecks.UI.Core", "src\HealthChecks.UI.Core\HealthChecks.UI.Core.csproj", "{146C2E30-B28B-45CA-A10C-6F3F1CF505F5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.UI.Client", "src\HealthChecks.UI.Client\HealthChecks.UI.Client.csproj", "{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -208,10 +211,6 @@ Global
 		{0A7AF6F0-DEA6-4195-91F3-CD9E2484E1F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0A7AF6F0-DEA6-4195-91F3-CD9E2484E1F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0A7AF6F0-DEA6-4195-91F3-CD9E2484E1F3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AA04C349-6402-4241-A611-60EA3CFBD4ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AA04C349-6402-4241-A611-60EA3CFBD4ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA04C349-6402-4241-A611-60EA3CFBD4ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -292,6 +291,10 @@ Global
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.Build.0 = Release|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -304,10 +307,18 @@ Global
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16}.Release|Any CPU.Build.0 = Release|Any CPU
+		{146C2E30-B28B-45CA-A10C-6F3F1CF505F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{146C2E30-B28B-45CA-A10C-6F3F1CF505F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{146C2E30-B28B-45CA-A10C-6F3F1CF505F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{146C2E30-B28B-45CA-A10C-6F3F1CF505F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -334,7 +345,6 @@ Global
 		{7C8F7A48-2FF9-4CF4-AAAC-142053E34FC0} = {092533AB-7505-4EDC-8932-D40BF575D0D2}
 		{EEF1E206-DEE3-45F2-A753-73D2FFFF8EDA} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{0A7AF6F0-DEA6-4195-91F3-CD9E2484E1F3} = {092533AB-7505-4EDC-8932-D40BF575D0D2}
-		{B1E1B671-EAA2-42A2-909E-0DE2C9900BF6} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{AA04C349-6402-4241-A611-60EA3CFBD4ED} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{79EA41F3-AEB2-49E0-9C13-E3741DAA3968} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{6FB1E70A-2915-4810-BCA4-AF38010AF949} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
@@ -355,10 +365,13 @@ Global
 		{18F9E412-646D-4751-9751-30AA7A0233DF} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{B526834E-9392-4749-BAB2-7DF579F8F418} = {092533AB-7505-4EDC-8932-D40BF575D0D2}
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{692313D3-E947-494A-83B7-754E2FFAF348} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{737E4FD6-EA77-4608-A20F-767557FE3190} = {95119F6F-87C8-45B8-8D95-61736FBEBEDE}
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{0CCC5412-3856-4DAF-BE7E-4AEF13DCCB16} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{146C2E30-B28B-45CA-A10C-6F3F1CF505F5} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{77CF612D-76FC-4F8A-8A30-2B61DC80DA5D} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2B8C62A1-11B6-469F-874C-A02443256568}
