+++ samples/HealthChecks.UIAndApiCustomization/Startup.cs
@@ -26,6 +26,7 @@ namespace HealthChecks.UIAndApi
                 //                        .AddWebhookNotification("webhook1", "http://webhook", "mypayload")
                 //                        .SetEvaluationTimeInSeconds(16);
                 //                })
+                .Services
                 .AddHealthChecks()
                 .AddUrlGroup(new Uri("http://httpbin.org/status/200"), name: "uri-1")
                 .AddUrlGroup(new Uri("http://httpbin.org/status/200"), name: "uri-2")
