+++ src/HealthChecks.UI.Client/UIResponseWriter.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using HealthChecks.UI.Core;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using System;
 using System.IO;
