+++ src/HealthChecks.UI.Core/Data/Configuration/HealthCheckConfigurationMap.cs
