+++ src/HealthChecks.UI.Core/Data/Configuration/HealthCheckExecutionEntryMap.cs
