+++ src/HealthChecks.UI.Core/Data/Configuration/HealthCheckExecutionHistoryMap.cs
