+++ src/HealthChecks.UI.Core/Data/Configuration/HealthCheckFailureNotificationsMap.cs
