+++ src/HealthChecks.UI.Core/Data/HealthCheckDbDesignTimeContextFactory.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Design;
+
+namespace HealthChecks.UI.Core.Data
+{
+    class HealthCheckDbDesignTimeContextFactory
+        
+    {
+        public HealthChecksDb CreateDbContext(string[] args)
+        {
+            var optionsBuilder = new DbContextOptionsBuilder<HealthChecksDb>();
+            //    .UseSqlite("Data Source=migrations");
+
+            return new HealthChecksDb(optionsBuilder.Options);
+        }
+    }
+}
