+++ src/HealthChecks.UI.Core/Data/HealthCheckExecutionEntry.cs
@@ -1,18 +1,17 @@
-﻿using HealthChecks.UI.Client;
-using System;
+﻿using System;
 
 namespace HealthChecks.UI.Core.Data
 {
-    public class HealthCheckExecutionHistory
+    public class HealthCheckExecutionEntry
     {
         public int Id { get; set; }
 
         public string Name { get; set; }
 
-        public string Description { get; set; }
-
         public UIHealthStatus Status { get; set; }
 
-        public DateTime On { get; set; }
+        public string Description { get; set; }
+
+        public TimeSpan Duration { get; set; }
     }
 }
