+++ src/HealthChecks.UI.Core/Data/HealthCheckFailureNotifications.cs
