+++ src/HealthChecks.UI.Core/HealthChecksUIBuilder.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public class HealthChecksUIBuilder
+    {
+        public IServiceCollection Services { get; }
+        public HealthChecksUIBuilder(IServiceCollection services)
+        {
+            Services = services ?? throw new ArgumentNullException(nameof(services));
+        }
+    }
+}
