+++ src/HealthChecks.UI.SqlServer.Storage/.config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "dotnet-ef": {
+      "version": "3.1.3",
+      "commands": [
+        "dotnet-ef"
+      ]
+    }
+  }
+}
