+++ src/HealthChecks.UI/Configuration/Settings.cs
@@ -1,4 +1,4 @@
-﻿using HealthChecks.UI.Client;
+﻿using HealthChecks.UI.Core;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
