--- src/HealthChecks.UI/Core/Data/HealthCheckExecutionEntry.cs
@@ -1,18 +1,0 @@
-﻿using HealthChecks.UI.Client;
-using System;
-
-namespace HealthChecks.UI.Core.Data
-{
-    public class HealthCheckExecutionEntry
-    {
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        public UIHealthStatus Status { get; set; }
-
-        public string Description { get; set; }
-
-        public TimeSpan Duration { get; set; }
-    }
-}
