+++ src/HealthChecks.UI/Core/Extensions/HealthReportExtensions.cs
@@ -1,4 +1,4 @@
-﻿using HealthChecks.UI.Client;
+﻿using HealthChecks.UI.Core;
 using HealthChecks.UI.Core.Data;
 using System.Collections.Generic;
 using System.Linq;
