+++ src/HealthChecks.UI/Endpoints/HealthChecksUIConventionBuilder.cs
