+++ src/HealthChecks.UI/Extensions/ApplicationBuilderExtensions.cs
