+++ src/HealthChecks.UI/Extensions/ConfigurationExtensions.cs
