+++ src/HealthChecks.UI/Extensions/EndpointRouteBuilderExtensions.cs
